requirements.txt
asgiref==3.8.1
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.4.4
Django==5.0.3
django-admin==2.0.2
django-autoslug==1.9.9
django-excel-response2==3.0.6
django-money==3.4.1
django-phonenumber-field==7.3.0
django-six==1.0.5
excel-base==1.0.4
idna==3.7
isoweek==1.3.3
phonenumberslite==8.13.33
pillow==10.2.0
psycopg2-binary==2.9.9
py-moneyed==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
requests==2.31.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
stripe==9.1.0
TimeConvert==3.0.13
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
xlwt==1.3.0
